create SearchOrOpenApiDescriptionCommand
